+++ test/dummy/test/test_helper.rb
@@ -4,3 +4,8 @@ ENV["RAILS_ENV"] ||= "test"
 
 require_relative "../config/environment"
 require "rails/test_help"
+require_relative "support/accessible_test_helpers"
+
+class ActionDispatch::IntegrationTest
+  include AccessibleTestHelpers
+end